grpc-gen-python/example/python/pyproject.toml
[tool.poetry]
name = "grpc-gen-python"
version = "0.1.0"
description = ""
authors = ["Osorio Rodrigues <{ID}+{username}@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
